// packages/engine/src/Space/Models/TSpaceServices.ts
import type { TActorService } from '@/Actor';
import type { TAnimationsService } from '@/Animations';
import type { TAudioService } from '@/Audio';
import type { TCameraService } from '@/Camera';
import type { TCollisionsService } from '@/Collisions';
import type { TControlsService } from '@/Controls';
import type { TEnvMapService } from '@/EnvMap';
import type { TFogService } from '@/Fog';
import type { TFsmService } from '@/Fsm';
import type { TIntersectionsWatcherService } from '@/Intersections';
import type { TKeyboardService } from '@/Keyboard';
import type { TLightService } from '@/Light';
import type { TLoopService } from '@/Loop';
import type { TMaterialService } from '@/Material';
import type { TModels3dService } from '@/Models3d';
import type { TMouseService } from '@/Mouse';
import type { TParticlesService } from '@/Particles';
import type { TPhysicsBodyService, TPhysicsWorldService } from '@/Physics';
import type { TRendererService } from '@/Renderer';
import type { TScenesService } from '@/Scene';
import type { TSpatialGridService } from '@/Spatial';
import type { TTextService } from '@/Text';
import type { TTextureService } from '@/Texture';
import type { TTransformDriveService } from '@/TransformDrive';

export type TSpaceServices = Readonly<{
  actorService: TActorService;
  animationsService: TAnimationsService;
  audioService: TAudioService;
  cameraService: TCameraService;
  collisionsService: TCollisionsService;
  controlsService: TControlsService;
  envMapService: TEnvMapService;
  fogService: TFogService;
  fsmService: TFsmService;
  intersectionsWatcherService: TIntersectionsWatcherService;
  keyboardService: TKeyboardService;
  lightService: TLightService;
  loopService: TLoopService;
  materialService: TMaterialService;
  models3dService: TModels3dService;
  mouseService: TMouseService;
  particlesService: TParticlesService;
  physicsBodyService: TPhysicsBodyService;
  physicsWorldService: TPhysicsWorldService;
  rendererService: TRendererService;
  scenesService: TScenesService;
  spatialGridService: TSpatialGridService;
  textService: TTextService;
  textureService: TTextureService;
  transformDriveService: TTransformDriveService;
}>;